Split out WUndergroundData to separate file
Because of the split, the base configuration (API key and PWS ID) needs to be
placed in a separate section of the `configuration.yaml`, outside of the
`sensors` section:

	wundergroundpws:
	  api_key: YOUR_API_KEY
	  pws_id: YOUR_STATION_ID
	  numeric_precision: none

	sensor:
	  - platform: wundergroundpws
	    monitored_conditions:
	      - temp
